cooldown + more embeds

+++ src/main/java/org/igsq/igsqbot/commands/fun/AvatarCommand.java
@@ -82,7 +82,7 @@ public class AvatarCommand extends Command
 	}
 
 	@Override
-	public int getCooldown()
+	public long getCooldown()
 	{
 		return 0;
 	}
