+++ src/main/java/org/igsq/igsqbot/commands/fun/MockCommand.java
@@ -80,7 +80,7 @@ public class MockCommand extends Command
 	}
 
 	@Override
-	public int getCooldown()
+	public long getCooldown()
 	{
 		return 0;
 	}
