+++ src/main/java/org/igsq/igsqbot/commands/misc/LinkCommand.java
@@ -179,7 +179,7 @@ public class LinkCommand extends Command
 	}
 
 	@Override
-	public int getCooldown()
+	public long getCooldown()
 	{
 		return 0;
 	}
