+++ src/main/java/org/igsq/igsqbot/commands/misc/SuggestionCommand.java
@@ -81,7 +81,7 @@ public class SuggestionCommand extends Command
 	}
 
 	@Override
-	public int getCooldown()
+	public long getCooldown()
 	{
 		return 0;
 	}
